% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_is_odd}
\alias{cpp_is_odd}
\title{Determine if integer is odd or not}
\usage{
cpp_is_odd(value = 0L)
}
\arguments{
\item{value}{Integer value}
}
\value{
boolean of whether value is odd
}
\description{
Determine if integer is odd or not
}
