\name{odder-package}
\alias{odder-package}
\alias{odder}
\docType{package}
\title{
  Robbie's odd package
}
\description{
  A package that has two ways of checking if an integer is odd
}
\details{
  There are two functions available for us to use here, R_is_odd, and cpp_is_odd, each of this take a an integer and return a boolean of whether that number is odd or not
}
\author{
Robert Davies

Maintainer: Robert Davies <robertwilliamdavies@gmail.com>
}
\examples{
  \dontrun{
     ## R_is_odd(10L)
     ## cpp_is_odd(10L)
  }
}
