% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odd.R
\name{R_is_odd}
\alias{R_is_odd}
\title{Determine in R if an integer is odd}
\usage{
R_is_odd(value = 0L)
}
\arguments{
\item{value}{An integer}
}
\value{
TRUE/FALSE of Whether the integer is odd
}
\description{
Determine in R if an integer is odd
}
\examples{
R_is_odd(10L)
R_is_odd(11L)
}
